created folders, moved tests accordingly
